feat(examples): split core sdk examples & add readme
