+++ examples/core/.env.example
@@ -1,0 +1,27 @@
+# Starknet RPC URL
+STARKNET_RPC_URL=http://0.0.0.0:5050
+# Solis RPC URL
+ARKCHAIN_RPC_URL=http://0.0.0.0:7777
+
+# Starknet accounts
+STARKNET_ACCOUNT1_ADDRESS=0xb3ff441a68610b30fd5e2abbf3a1548eb6ba6f3559f2862bf2dc757e5828ca
+STARKNET_ACCOUNT1_PRIVATE_KEY=0x2bbf4f9fd0bbb2e60b0316c1fe0b76cf7a4d0198bd493ced9b8df2a3a24d68a
+STARKNET_ACCOUNT1_PUBLIC_KEY=0x4c339f18b9d1b95b64a6d378abd1480b2e0d5d5bd33cd0828cbce4d65c27284
+
+STARKNET_ACCOUNT2_ADDRESS=0xe29882a1fcba1e7e10cad46212257fea5c752a4f9b1b1ec683c503a2cf5c8a
+STARKNET_ACCOUNT2_PRIVATE_KEY=0x14d6672dcb4b77ca36a887e9a11cd9d637d5012468175829e9c6e770c61642
+STARKNET_ACCOUNT2_PUBLIC_KEY=0x4c0f884b8e5b4f00d97a3aad26b2e5de0c0c76a555060c837da2e287403c01d
+
+SOLIS_ADMIN_ADDRESS_DEV=0x6b86e40118f29ebe393a75469b4d926c7a44c2e2681b6d319520b7c1156d114
+SOLIS_ADMIN_PRIVATE_KEY_DEV=0x1c9053c053edf324aec366a34c6901b1095b07af69495bffec7d7fe21effb1b
+SOLIS_ADMIN_PUBLIC_KEY_DEV=0x2b191c2f3ecf685a91af7cf72a43e7b90e2e41220175de5c4f7498981b10053
+
+SOLIS_ADMIN_ADDRESS=0x6b86e40118f29ebe393a75469b4d926c7a44c2e2681b6d319520b7c1156d114
+SOLIS_ADMIN_PRIVATE_KEY=0x1c9053c053edf324aec366a34c6901b1095b07af69495bffec7d7fe21effb1b
+
+# Starknet network
+STARKNET_NETWORK_ID=dev
+# Solis network
+SOLIS_NETWORK_ID=dev
+
+BROKER_ID=0x6b86e40118f29ebe393a75469b4d926c7a44c2e2681b6d319520b7c1156d114
