+++ examples/core/cancelAuction.ts
@@ -2,18 +2,16 @@ import "dotenv/config";
 
 import { stark } from "starknet";
 
-import { createBroker } from "../src/actions/broker/createBroker.js";
 import {
   approveERC721,
   AuctionV1,
+  cancelOrder,
   createAuction,
-  createOffer,
+  createBroker,
   fetchOrCreateAccount,
-  fulfillAuction,
-  getOrderStatus,
-  OfferV1
-} from "../src/index.js";
-import { FulfillAuctionInfo } from "../src/types/index.js";
+  getOrderStatus
+} from "@ark-project/core";
+
 import { config } from "./config/index.js";
 import { STARKNET_NFT_ADDRESS } from "./constants/index.js";
 import { getCurrentTokenId } from "./utils/getCurrentTokenId.js";
@@ -24,20 +22,13 @@ import { whitelistBroker } from "./utils/whitelistBroker.js";
   // Create test accounts
   const adminAccount = await fetchOrCreateAccount(
     config.arkProvider,
-    process.env.SOLIS_ADMIN_ADDRESS,
-    process.env.SOLIS_ADMIN_PRIVATE_KEY
+    process.env.SOLIS_ADMIN_ADDRESS_DEV,
+    process.env.SOLIS_ADMIN_PRIVATE_KEY_DEV
   );
-
   const sellerAccount = await fetchOrCreateAccount(
     config.starknetProvider,
     process.env.STARKNET_ACCOUNT1_ADDRESS,
     process.env.STARKNET_ACCOUNT1_PRIVATE_KEY
-  );
-
-  const buyerAccount = await fetchOrCreateAccount(
-    config.starknetProvider,
-    process.env.STARKNET_ACCOUNT2_ADDRESS,
-    process.env.STARKNET_ACCOUNT2_PRIVATE_KEY
   );
 
   // Create and whitelist broker
@@ -70,38 +61,22 @@ import { whitelistBroker } from "./utils/whitelistBroker.js";
     order
   });
 
-  // Create offer
-  const offer: OfferV1 = {
-    brokerId,
-    tokenAddress: STARKNET_NFT_ADDRESS,
-    tokenId,
-    startAmount: 1
-  };
-
-  const offerOrderHash = await createOffer(config, {
-    starknetAccount: buyerAccount,
-    arkAccount: adminAccount,
-    offer
-  });
-
-  // Fulfill auction
-  const fulfillAuctionInfo: FulfillAuctionInfo = {
-    orderHash,
-    relatedOrderHash: offerOrderHash,
+  // Cancel auction
+  const cancelInfo = {
+    orderHash: orderHash,
     tokenAddress: order.tokenAddress,
-    tokenId,
-    brokerId
+    tokenId: order.tokenId
   };
 
-  await fulfillAuction(config, {
+  cancelOrder(config, {
     starknetAccount: sellerAccount,
     arkAccount: adminAccount,
-    fulfillAuctionInfo
+    cancelInfo
   });
 
   await new Promise((resolve) => setTimeout(resolve, 10_000));
 
-  const { orderStatus: orderStatus } = await getOrderStatus(config, {
+  const { orderStatus } = await getOrderStatus(config, {
     orderHash
   });
