+++ examples/core/createAuction.ts
@@ -2,13 +2,14 @@ import "dotenv/config";
 
 import { stark } from "starknet";
 
-import { createBroker } from "../src/actions/broker/createBroker.js";
 import {
   approveERC721,
   createAuction,
+  createBroker,
   fetchOrCreateAccount,
   getOrderStatus
-} from "../src/index.js";
+} from "@ark-project/core";
+
 import { config } from "./config/index.js";
 import { STARKNET_NFT_ADDRESS } from "./constants/index.js";
 import { getCurrentTokenId } from "./utils/getCurrentTokenId.js";
