+++ examples/core/fulfillListing.ts
@@ -4,21 +4,22 @@
  * submitting a listing order and cancelling it.
  */
 
-import { stark } from "starknet";
-
 import "dotenv/config";
 
-import { createBroker } from "../src/actions/broker/createBroker.js";
+import { stark } from "starknet";
+
 import {
   approveERC20,
   approveERC721,
   createAccount,
-  createOffer,
+  createBroker,
+  createListing,
   fetchOrCreateAccount,
-  fulfillOffer,
+  fulfillListing,
   getOrderStatus,
-  OfferV1
-} from "../src/index.js";
+  ListingV1
+} from "@ark-project/core";
+
 import { config } from "./config/index.js";
 import {
   STARKNET_ETH_ADDRESS,
@@ -33,8 +34,6 @@ import { whitelistBroker } from "./utils/whitelistBroker.js";
  * Creates a listing on the blockchain using provided order details.
  */
 (async () => {
-  console.log(`=> Getting config...`);
-  const { arkProvider, starknetProvider } = config;
   const brokerId = stark.randomAddress();
   await createBroker(config, { brokerID: brokerId });
 
@@ -44,43 +43,14 @@ import { whitelistBroker } from "./utils/whitelistBroker.js";
     process.env.SOLIS_ADMIN_PRIVATE_KEY_DEV
   );
 
+  console.log(`=> Whitelisting broker ${brokerId}`);
   await whitelistBroker(config, solisAdminAccount, brokerId);
 
-  console.log(`=> Creating account`);
+  console.log(`=> Getting config...`);
+  const { arkProvider, starknetProvider } = config;
+  console.log(`=> Creating ark account`);
   // Create a new account for the listing using the provider
   const { account: arkAccount } = await createAccount(arkProvider);
-
-  console.log(
-    `=> Fetching or creating fulfiller starknet account, for test purpose only`
-  );
-  const starknetFulfillerAccount = await fetchOrCreateAccount(
-    starknetProvider,
-    process.env.STARKNET_ACCOUNT2_ADDRESS,
-    process.env.STARKNET_ACCOUNT2_PRIVATE_KEY
-  );
-
-  console.log(
-    `=> Minting token at contract address: ${STARKNET_NFT_ADDRESS} with fulfiller account: ${starknetFulfillerAccount.address}`
-  );
-  await mintERC721(starknetProvider, starknetFulfillerAccount);
-  if (config.starknetNetwork !== "dev") {
-    console.log(
-      "=> Waiting for 5 minutes for transaction complete on goerli..."
-    );
-    await new Promise((resolve) => setTimeout(resolve, 5 * 60 * 1000));
-  }
-
-  const tokenId = await getCurrentTokenId(config, STARKNET_NFT_ADDRESS);
-  console.log("=> Token minted with tokenId: ", tokenId);
-
-  console.log(`=> Creating offer for tokenId: ${tokenId}`);
-  // Define the order details
-  const offer: OfferV1 = {
-    brokerId, // The broker ID
-    tokenAddress: STARKNET_NFT_ADDRESS, // The token address
-    tokenId: tokenId, // The ID of the token
-    startAmount: 100000000000000000 // The starting amount for the order
-  };
 
   console.log(
     `=> Fetching or creating offerer starknet account, for test purpose only`
@@ -91,38 +61,13 @@ import { whitelistBroker } from "./utils/whitelistBroker.js";
     process.env.STARKNET_ACCOUNT1_PRIVATE_KEY
   );
 
-  if (process.env.STARKNET_NETWORK_ID === "dev") {
-    console.log("=> Minting ERC20...");
-    await mintERC20(
-      starknetProvider,
-      starknetOffererAccount,
-      offer.startAmount
-    );
-  }
-
-  console.log(
-    `=> Approuving ERC20 tokens ${STARKNET_ETH_ADDRESS} from minter: ${starknetOffererAccount.address} to ArkProject executor`
+  console.log("=> Minting token at contract address: ", STARKNET_NFT_ADDRESS);
+  const transaction_hash = await mintERC721(
+    starknetProvider,
+    starknetOffererAccount
   );
-  await approveERC20(config, {
-    starknetAccount: starknetOffererAccount,
-    contractAddress: STARKNET_ETH_ADDRESS,
-    amount: offer.startAmount
-  });
 
-  console.log("=> Creating offer...");
-  // Create the listing on the arkchain using the order details
-  const orderHash = await createOffer(config, {
-    starknetAccount: starknetOffererAccount,
-    arkAccount,
-    offer
-  });
-
-  console.log(`=> Approving collection ${offer.tokenId}`);
-  await approveERC721(config, {
-    contractAddress: STARKNET_NFT_ADDRESS,
-    starknetAccount: starknetFulfillerAccount,
-    tokenId
-  });
+  console.log(transaction_hash);
 
   if (config.starknetNetwork !== "dev") {
     console.log(
@@ -131,20 +76,82 @@ import { whitelistBroker } from "./utils/whitelistBroker.js";
     await new Promise((resolve) => setTimeout(resolve, 5 * 60 * 1000));
   }
 
+  const tokenId = await getCurrentTokenId(config, STARKNET_NFT_ADDRESS);
+  console.log("=> Token minted with tokenId: ", tokenId);
+
+  console.log(`=> Approving for all`);
+  await approveERC721(config, {
+    contractAddress: STARKNET_NFT_ADDRESS,
+    starknetAccount: starknetOffererAccount,
+    tokenId
+  });
+
+  console.log(`=> Creating order`);
+  // Define the order details
+  const order: ListingV1 = {
+    brokerId, // The broker ID
+    tokenAddress: STARKNET_NFT_ADDRESS, // The token address
+    tokenId: tokenId, // The ID of the token
+    startAmount: 100000000000000000 // The starting amount for the order
+  };
+
+  console.log("=> Creating listing...");
+  // Create the listing on the arkchain using the order details
+  const orderHash = await createListing(config, {
+    starknetAccount: starknetOffererAccount,
+    arkAccount,
+    order
+  });
+
+  if (config.starknetNetwork !== "dev") {
+    await new Promise((resolve) => setTimeout(resolve, 5000));
+  }
+  console.log(
+    `=> Fetching or creating fulfiller starknet account, for test purpose only`
+  );
+  const starknetFulfillerAccount = await fetchOrCreateAccount(
+    starknetProvider,
+    process.env.STARKNET_ACCOUNT2_ADDRESS,
+    process.env.STARKNET_ACCOUNT2_PRIVATE_KEY
+  );
+
+  if (process.env.STARKNET_NETWORK_ID === "dev") {
+    console.log("=> Minting ERC20...");
+    await mintERC20(
+      starknetProvider,
+      starknetFulfillerAccount,
+      order.startAmount
+    );
+  }
+
+  console.log(
+    `=> Approving ERC20 tokens ${STARKNET_ETH_ADDRESS} from minter: ${starknetFulfillerAccount.address} to ArkProject executor`
+  );
+
+  await approveERC20(config, {
+    starknetAccount: starknetFulfillerAccount,
+    contractAddress: STARKNET_ETH_ADDRESS,
+    amount: order.startAmount
+  });
+
+  if (config.starknetNetwork !== "dev") {
+    await new Promise((resolve) => setTimeout(resolve, 2000));
+  }
+  console.log("tokenId", tokenId);
   // Define the fulfill details
-  const fulfillOfferInfo = {
+  const fulfillListingInfo = {
     orderHash: orderHash,
-    tokenAddress: offer.tokenAddress,
-    tokenId: offer.tokenId,
+    tokenAddress: order.tokenAddress,
+    tokenId,
     brokerId
   };
 
-  console.log(`=> Fulfilling offer by ${starknetFulfillerAccount.address}`);
+  console.log(`=> Fulfilling listing by ${starknetFulfillerAccount.address}`);
   // fulfill the order
-  await fulfillOffer(config, {
+  await fulfillListing(config, {
     starknetAccount: starknetFulfillerAccount,
     arkAccount,
-    fulfillOfferInfo
+    fulfillListingInfo
   });
 
   if (config.starknetNetwork !== "dev") {
@@ -153,7 +160,7 @@ import { whitelistBroker } from "./utils/whitelistBroker.js";
     );
     await new Promise((resolve) => setTimeout(resolve, 5 * 60 * 1000));
   } else {
-    await new Promise((resolve) => setTimeout(resolve, 10000));
+    await new Promise((resolve) => setTimeout(resolve, 5000));
   }
 
   console.log("=> Fetching order status...");
