+++ examples/core/tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "extends": "@ark-project/typescript-config/base.json",
+  "compilerOptions": {
+    "target": "esnext",
+    "plugins": [
+      {}
+    ]
+  },
+  "include": [
+    "**/*.ts",
+  ],
+  "exclude": [
+    "node_modules"
+  ]
+}
