+++ examples/core/utils/getBalance.ts
