+++ examples/core/utils/getCurrentTokenId.ts
