+++ examples/core/utils/getTokenOwner.ts
