+++ examples/core/utils/mintERC20.ts
