+++ examples/core/utils/setArkFees.ts
