+++ packages/core/package.json
@@ -5,7 +5,6 @@
   "scripts": {
     "build": "npm run generate-contracts && pnpm run clean && pnpm run build:esm+types",
     "build:esm+types": "tsc --project tsconfig.build.json --outDir ./dist/esm --declaration --declarationMap --declarationDir ./dist/types",
-    "build:examples": "tsc --project tsconfig.json --outDir ./dist --declaration --declarationMap --declarationDir ./dist/types",
     "clean": "rm -rf dist tsconfig.tsbuildinfo actions chains codegen internal query",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
@@ -15,13 +14,13 @@
     "test": "jest tests -i"
   },
   "type": "module",
-  "main": "./dist/esm/src/index.js",
-  "types": "./dist/types/src/index.d.ts",
-  "typings": "./dist/types/src/index.d.ts",
+  "main": "./dist/esm/index.js",
+  "types": "./dist/types/index.d.ts",
+  "typings": "./dist/types/index.d.ts",
   "exports": {
     ".": {
-      "types": "./dist/types/src/index.d.ts",
-      "default": "./dist/esm/src/index.js"
+      "types": "./dist/types/index.d.ts",
+      "default": "./dist/esm/index.js"
     }
   },
   "dependencies": {
