+++ packages/core/tsconfig.build.json
@@ -2,7 +2,6 @@
   "extends": "@ark-project/typescript-config/base.json",
   "include": [
     "src/**/*.ts",
-    "examples/**/*.ts"
   ],
   "exclude": [
     "**/*.test.ts",
